fix dependency issue on travis
