+++ webapp/model/Lifespan.php
@@ -1,15 +1,5 @@
 <?php
-
-interface LifespanInterface {
-
-    public function __construct($lifespanID);
-    public function status();
-    public function isCurrent();
-    public function offered();
-    public function accepted();
-    public function declined();
-    public function isEnded();
-}
+require 'LifespanInterface.php';
 
 class Lifespan implements LifespanInterface {
